feat: port intrerrupt node to planner
