+++ apps/open-swe/src/graphs/planner/nodes/index.ts
@@ -2,3 +2,5 @@ export * from "./generate-message/index.js";
 export * from "./take-action.js";
 export * from "./generate-plan.js";
 export * from "./summarizer.js";
+export * from "./proposed-plan.js";
+export * from "./prepare-state.js";
