+++ apps/open-swe/src/utils/user-request.ts
@@ -17,7 +17,9 @@ export function getUserRequest(
   messages: BaseMessage[],
   options?: { returnFullMessage?: boolean },
 ): string | HumanMessage {
-  const recentUserMessage = messages.findLast(isHumanMessage);
+  const recentUserMessage = messages.findLast(
+    (m) => isHumanMessage(m) && m.additional_kwargs?.isOriginalIssue,
+  );
   if (!recentUserMessage) {
     return "";
   }
